+++ tests/testthat/test_exact_rank_prob.R
@@ -1,7 +1,7 @@
 context("probabilistic centrality")
 library(igraph)
 library(magrittr)
-
+library(Matrix)
 test_that("exact_rank_prob is correct",{
   P <- matrix(c(0,0,1,1,1,0,0,0,1,0,0,0,0,0,1,rep(0,10)),5,5,byrow=TRUE)
   res <- exact_rank_prob(P)