+++ tests/testthat/test_indirect_relations.R
@@ -1,5 +1,7 @@
 context("indirect relations")
 library(igraph)
+library(magrittr)
+library(Matrix)
 
 test_that("adjacency is correct",{
   g <- graph.empty(n=11,directed = FALSE)
@@ -10,8 +12,7 @@ test_that("adjacency is correct",{
 })
 
 test_that("shortest path distances are correct",{
-  # library(igraph)
-  
+
   g <- graph.empty(n=11,directed = FALSE)
   g <- add_edges(g,c(1,11,2,4,3,5,3,11,4,8,5,9,5,11,6,7,6,8,
                      6,10,6,11,7,9,7,10,7,11,8,9,8,10,9,10))
@@ -35,8 +36,6 @@ test_that("shortest path distances are correct",{
 
 
 test_that("walk counts are correct",{
-  # library(igraph)
-  
   g <- graph.empty(n=11,directed = FALSE)
   g <- add_edges(g,c(1,11,2,4,3,5,3,11,4,8,5,9,5,11,6,7,6,8,
                      6,10,6,11,7,9,7,10,7,11,8,9,8,10,9,10))
@@ -97,11 +96,11 @@ test_that("walk transform is correct",{
   expect_equal(A,matrix(4,5,5))
   
   katz <- rowSums(indirect_relations(g,type = "walks", FUN=walks_attenuated,alpha=0.1))
-  testthat::expect_equal(katz,rep(5/3,5))
+  expect_equal(katz,rep(5/3,5))
   
   even_and_odd <- indirect_relations(g,type = "walks", FUN=walks_exp_even)+
     indirect_relations(g,type = "walks", FUN=walks_exp_odd)
   all <- indirect_relations(g,type = "walks", FUN=walks_exp)
-  testthat::expect_equal(even_and_odd,all)
+  expect_equal(even_and_odd,all)
   
 })